-  fix now shown enable auto sync snackbar  - update subs dialog title  - update login description

+++ scp2/src/main/java/ru/dante/scpfoundation/manager/MyPreferenceManager.java
@@ -229,12 +229,12 @@ public class MyPreferenceManager {
     }
 
     //auto sync
-    public void setNumOfAttemptsToAutoSync(int numOfAttemptsToAutoSync) {
-        mPreferences.edit().putInt(Keys.AUTO_SYNC_ATTEMPTS, numOfAttemptsToAutoSync).apply();
+    public void setNumOfAttemptsToAutoSync(long numOfAttemptsToAutoSync) {
+        mPreferences.edit().putLong(Keys.AUTO_SYNC_ATTEMPTS, numOfAttemptsToAutoSync).apply();
     }
 
-    public int getNumOfAttemptsToAutoSync() {
-        return mPreferences.getInt(Keys.AUTO_SYNC_ATTEMPTS, 0);
+    public long getNumOfAttemptsToAutoSync() {
+        return mPreferences.getLong(Keys.AUTO_SYNC_ATTEMPTS, 0);
     }
 
     //secure
